+++ packages/sampler/stories/qa/Spotlight.js
@@ -410,7 +410,7 @@ storiesOf('Spotlight', module)
 					Use the knobs to test the available behaviors for the spottable components
 					below.
 				</p>
-				<Container style={style.flexBox} spotlightMuted={boolean('spotlightMuted', Container, false)}>
+				<Container style={style.flexBox} spotlightMuted={boolean('spotlightMuted', Container, false)} spotlightDisabled={boolean('Container spotlightDisabled', Container, false)}>
 					<div style={style.flexItem}>
 						<Divider>
 							Misc Components
@@ -421,7 +421,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							>
 								Button
 							</Button>
@@ -431,7 +431,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							>
 								Translucent
 							</Button>
@@ -443,7 +443,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							>
 								Transparent
 							</Button>
@@ -452,7 +452,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							>
 								ToggleButton
 							</ToggleButton>
@@ -463,7 +463,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							>
 								plus
 							</IconButton>
@@ -472,7 +472,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							/>
 						</div>
 						<div style={style.flexBox}>
@@ -481,7 +481,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							>
 								{Items}
 							</Picker>
@@ -491,7 +491,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 							>
 								{Items}
 							</Picker>
@@ -501,21 +501,21 @@ storiesOf('Spotlight', module)
 							onSpotlightLeft={action('onSpotlightLeft')}
 							onSpotlightRight={action('onSpotlightRight')}
 							onSpotlightUp={action('onSpotlightUp')}
-							spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+							spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 						/>
 						<Slider
 							onSpotlightDown={action('onSpotlightDown')}
 							onSpotlightLeft={action('onSpotlightLeft')}
 							onSpotlightRight={action('onSpotlightRight')}
 							onSpotlightUp={action('onSpotlightUp')}
-							spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+							spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 						/>
 						<Item
 							onSpotlightDown={action('onSpotlightDown')}
 							onSpotlightLeft={action('onSpotlightLeft')}
 							onSpotlightRight={action('onSpotlightRight')}
 							onSpotlightUp={action('onSpotlightUp')}
-							spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+							spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 						>
 							Item
 						</Item>
@@ -525,7 +525,7 @@ storiesOf('Spotlight', module)
 							onSpotlightLeft={action('onSpotlightLeft')}
 							onSpotlightRight={action('onSpotlightRight')}
 							onSpotlightUp={action('onSpotlightUp')}
-							spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+							spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 						>
 							LabeledItem
 						</LabeledItem>
@@ -540,7 +540,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								title="Various Items in an ExpandableItem"
 							>
 								<CheckboxItem
@@ -548,7 +548,7 @@ storiesOf('Spotlight', module)
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								>
 									CheckboxItem
 								</CheckboxItem>
@@ -557,7 +557,7 @@ storiesOf('Spotlight', module)
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								>
 									FormCheckboxItem
 								</FormCheckboxItem>
@@ -566,7 +566,7 @@ storiesOf('Spotlight', module)
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								>
 									RadioItem
 								</RadioItem>
@@ -575,7 +575,7 @@ storiesOf('Spotlight', module)
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								>
 									SelectableItem
 								</SelectableItem>
@@ -584,7 +584,7 @@ storiesOf('Spotlight', module)
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								>
 									SwitchItem
 								</SwitchItem>
@@ -595,7 +595,7 @@ storiesOf('Spotlight', module)
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								>
 									ToggleItem
 								</ToggleItem>
@@ -606,7 +606,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								title="ExpandableList"
 							>
 								{Items}
@@ -616,7 +616,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								title="ExpandableInput"
 							/>
 							<ExpandablePicker
@@ -624,7 +624,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								title="ExpandablePicker"
 							>
 								{Items}
@@ -634,7 +634,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								title="DatePicker"
 							/>
 							<DayPicker
@@ -642,7 +642,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								title="DayPicker"
 							/>
 							<TimePicker
@@ -650,7 +650,7 @@ storiesOf('Spotlight', module)
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('spotlightDisabled', Container, false)}
+								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
 								title="TimePicker"
 							/>
 						</Scroller>